Update project configuration and add browser support
