+++ .npmignore
@@ -3,8 +3,8 @@ bin/guida.js
 elm-stuff
 examples
 guida-stuff
-lib
 review
+scripts
 src
 tests
 .nvmrc
