+++ .nvmrc
@@ -1,1 +1,1 @@
-v23.7.0
+v23.10.0
