--- browser/public/index.html
@@ -1,13 +1,0 @@
-<!doctype html>
-<html>
-
-<head>
-    <meta charset="UTF-8" />
-    <title>Guida</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script src="/app.js"></script>
-</head>
-
-<body></body>
-
-</html>
