+++ package.json
@@ -4,20 +4,18 @@
   "description": "Guida is a functional programming language that builds upon the solid foundation of Elm, offering backward compatibility with all existing Elm 0.19.1 projects",
   "author": "Decio Ferreira",
   "license": "BSD-3-Clause",
-  "main": "./lib/guida.js",
-  "bin": {
-    "guida": "bin/index.js"
-  },
+  "browser": "lib/index.js",
+  "bin": "bin/index.js",
   "scripts": {
-    "build": "./lib/guida.sh src/Terminal/Main.elm",
+    "build": "npm-run-all --parallel build:*",
+    "build:bin": "./scripts/build.sh bin",
+    "build:browser": "./scripts/build.sh browser",
     "test": "npm-run-all --parallel test:*",
     "test:jest": "jest",
     "test:elm": "elm-test",
     "test:elm-review": "elm-review",
     "test:elm-format-validate": "elm-format . --validate",
-    "elm-format": "elm-format . --yes",
-    "prebrowser": "esbuild browser/index.js --bundle --outfile=browser/public/app.js",
-    "browser": "./browser/server.js"
+    "elm-format": "elm-format . --yes"
   },
   "dependencies": {
     "adm-zip": "^0.5.16",
@@ -29,15 +27,11 @@
     "which": "^5.0.0"
   },
   "devDependencies": {
-    "cors": "^2.8.5",
     "elm": "^0.19.1-6",
     "elm-format": "^0.8.7",
     "elm-review": "^2.13.1",
     "elm-test": "^0.19.1-revision12",
-    "esbuild": "^0.25.1",
-    "express": "^4.21.2",
     "guida": "^0.3.0-alpha",
-    "http-proxy-middleware": "^3.0.3",
     "jest": "^29.7.0",
     "npm-run-all": "^4.1.5",
     "uglify-js": "^3.19.3"
