+++ scripts/build.sh
@@ -4,10 +4,25 @@
 
 set -e
 
-js="bin/guida.js"
-min="bin/guida.min.js"
+case $1 in
+  "bin")
+    dir="bin"
+    elm_entry="src/Terminal/Main.elm"
+    ;;
+  "browser")
+    dir="lib"
+    elm_entry="src/Browser/Main.elm"
+    ;;
+  *)
+    echo "Usage: $0 bin|browser"
+    exit 1
+    ;;
+esac
 
-guida make --optimize --output=$js $@
+js="$dir/guida.js"
+min="$dir/guida.min.js"
+
+elm make --optimize --output=$js $elm_entry
 
 uglifyjs $js --compress "pure_funcs=[F2,F3,F4,F5,F6,F7,F8,F9,A2,A3,A4,A5,A6,A7,A8,A9],pure_getters,keep_fargs=false,unsafe_comps,unsafe" | uglifyjs --mangle --output $min
