+++ try/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "guida-try",
+  "version": "1.0.0",
+  "main": "index.js",
+  "scripts": {
+    "preserver": "esbuild app.js --bundle --platform=browser --outfile=public/app.js",
+    "server": "./server.js"
+  },
+  "dependencies": {
+    "cors": "^2.8.5",
+    "esbuild": "^0.25.1",
+    "express": "^4.21.2",
+    "guida": "file:..",
+    "http-proxy-middleware": "^3.0.3"
+  }
+}
